fix(createPage): refactor for create client view
